+++ src/macos/macos_impl.rs
@@ -83,6 +83,8 @@ pub struct Enigo {
     // Instant when the last event was sent and the duration that needs to be waited for after that
     // instant to make sure all events were handled by the OS
     last_event: (Instant, Duration),
+    // The last location the mouse was programmatically moved to and then instant when it happened
+    last_mouse_move: (CGPoint, Instant),
     // TODO: Use mem::variant_count::<Button>() here instead of 9 once it is stabilized
     last_mouse_click: [(i64, Instant); 9], /* For each of the nine Button variants, we
                                             * store the last time the button was clicked and
@@ -100,7 +102,8 @@ impl Mouse for Enigo {
     // Sends a button event to the X11 server via `XTest` extension
     fn button(&mut self, button: Button, direction: Direction) -> InputResult<()> {
         debug!("\x1b[93mbutton(button: {button:?}, direction: {direction:?})\x1b[0m");
-        let (current_x, current_y) = self.location()?;
+
+        let dest = self.mouse_location()?;
 
         if direction == Direction::Click || direction == Direction::Press {
             let click_count = self.nth_button_press(button, Direction::Press);
@@ -115,7 +118,6 @@ impl Mouse for Enigo {
                 Button::ScrollLeft => return self.scroll(-1, Axis::Horizontal),
                 Button::ScrollRight => return self.scroll(1, Axis::Horizontal),
             };
-            let dest = CGPoint::new(current_x as f64, current_y as f64);
 
             let event =
                 CGEvent::new_mouse_event(Some(&self.event_source), event_type, dest, button);
@@ -145,6 +147,8 @@ impl Mouse for Enigo {
             );
 
             CGEvent::set_flags(event.as_deref(), self.event_flags);
+            // No need to do self.update_event_location(&event) because it gets created with
+            // the correct coordinates
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -166,7 +170,7 @@ impl Mouse for Enigo {
                     return Ok(());
                 }
             };
-            let dest = CGPoint::new(current_x as f64, current_y as f64);
+
             let event =
                 CGEvent::new_mouse_event(Some(&self.event_source), event_type, dest, button);
             if event.is_none() {
@@ -193,6 +197,8 @@ impl Mouse for Enigo {
                 self.event_source_user_data,
             );
             CGEvent::set_flags(event.as_deref(), self.event_flags);
+            // No need to do self.update_event_location(&event) because it gets created with
+            // the correct coordinates
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -201,6 +207,7 @@ impl Mouse for Enigo {
 
     fn move_mouse(&mut self, x: i32, y: i32, coordinate: Coordinate) -> InputResult<()> {
         debug!("\x1b[93mmove_mouse(x: {x:?}, y: {y:?}, coordinate:{coordinate:?})\x1b[0m");
+
         let pressed = NSEvent::pressedMouseButtons();
         let (current_x, current_y) = self.location()?;
 
@@ -246,7 +253,10 @@ impl Mouse for Enigo {
             self.event_source_user_data,
         );
         CGEvent::set_flags(event.as_deref(), self.event_flags);
+        // No need to do self.update_event_location(&event) because it gets created with
+        // the correct coordinates
         CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
+        self.last_mouse_move = (dest, Instant::now());
         self.update_wait_time();
         Ok(())
     }
@@ -272,9 +282,9 @@ impl Mouse for Enigo {
 
     fn location(&self) -> InputResult<(i32, i32)> {
         debug!("\x1b[93mlocation()\x1b[0m");
-        let pt = NSEvent::mouseLocation();
-        let (x, y_inv) = (pt.x as i32, pt.y as i32);
-        Ok((x, self.display.pixels_high() as i32 - y_inv))
+
+        let location = self.mouse_location()?;
+        Ok(location.x as i32, location.y as i32);
     }
 }
 
@@ -345,6 +355,7 @@ impl Keyboard for Enigo {
             );
             // We want to ignore all modifiers when entering text
             CGEvent::set_flags(event.as_deref(), CGEventFlags::empty());
+            // TODO: check if we have to do: self.update_event_location(&event);
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -481,6 +492,7 @@ impl Keyboard for Enigo {
             );
             self.add_event_flag(keycode, Direction::Press);
             CGEvent::set_flags(event.as_deref(), self.event_flags);
+            self.update_event_location(&event);
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -500,6 +512,7 @@ impl Keyboard for Enigo {
             );
             self.add_event_flag(keycode, Direction::Release);
             CGEvent::set_flags(event.as_deref(), self.event_flags);
+            self.update_event_location(&event);
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -565,6 +578,11 @@ impl Enigo {
         debug!("\x1b[93mconnection established on macOS\x1b[0m");
 
         let last_event = (Instant::now(), Duration::from_secs(0));
+
+        let event = CGEvent::new(event_source.clone())
+            .map_err(|()| NewConError::EstablishCon("failed to create CGEvent"))?;
+        let last_mouse_move = (event.location(), Instant::now());
+
         Ok(Enigo {
             event_source,
             display: CGDisplay::main(),
@@ -573,6 +591,7 @@ impl Enigo {
             event_flags,
             double_click_delay,
             last_event,
+            last_mouse_move,
             last_mouse_click: [(0, Instant::now()); 9],
             event_source_user_data: event_source_user_data.unwrap_or(crate::EVENT_MARKER as i64),
         })
@@ -637,6 +656,7 @@ impl Enigo {
                 self.event_source_user_data,
             );
             CGEvent::set_flags(event.as_deref(), self.event_flags);
+            self.update_event_location(&cg_event);
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -667,6 +687,7 @@ impl Enigo {
                 self.event_source_user_data,
             );
             CGEvent::set_flags(event.as_deref(), self.event_flags);
+            self.update_event_location(&cg_event);
             CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
             self.update_wait_time();
         }
@@ -934,9 +955,36 @@ impl Enigo {
             self.event_source_user_data,
         );
         CGEvent::set_flags(event.as_deref(), self.event_flags);
+        self.update_event_location(&event);
         CGEvent::post(CGEventTapLocation::HIDEventTap, event.as_deref());
         self.update_wait_time();
         Ok(())
+    }
+
+    // Moving the mouse programmatically takes a few milli seconds. In order to
+    // avoid sleeping until all events are processed, we internally keep track of
+    // where the mouse was last moved to programmatically and return that location
+    // if the mouse move happened within the last few milliseconds. If it has been
+    // more than that, we query the current mouse location. That is because the user
+    // might have moved the mouse since then
+    fn mouse_location(&self) -> InputResult<CGPoint> {
+        let last_time = self.last_mouse_move.1;
+        if last_time.elapsed() < Duration::from_millis(60) {
+            Ok(self.last_mouse_move.0)
+        } else {
+            let event = CGEvent::new(self.event_source.clone())
+                .map_err(|()| InputError::Simulate("failed to create CGEvent"))?;
+            Ok(event.location())
+        }
+    }
+
+    // Manually update the location of an event if the OS did not yet have time to
+    // handle a simulated mouse move
+    fn update_event_location(&mut self, event: &CGEvent) {
+        let last_time = self.last_mouse_move.1;
+        if last_time.elapsed() < Duration::from_millis(60) {
+            event.set_location(self.last_mouse_move.0);
+        }
     }
 
     /// Save the current Instant and calculate the remaining waiting time
@@ -1207,7 +1255,6 @@ impl Drop for Enigo {
         // This sleep is needed because all events that have not been
         // processed until this point would just get ignored when the
         // struct is dropped
-        self.update_wait_time();
-        thread::sleep(self.last_event.1.saturating_sub(Duration::from_millis(20)));
+        thread::sleep(self.last_event.1);
     }
 }