+++ CHANGES.md
@@ -12,6 +12,7 @@
 - linux: Added `NumpadEnter` key
 - win: Added `NumpadEnter` key
 - win: Extended keys can now be correctly simulated with the raw() function.
+- macos: `move_mouse` no longer needs a sleep.
 
 # 0.6.1
 ## Fixed